fix(core): make sure all options of a query are reactive (#7081)
* fix: make sure all options of a query are reactive

currently, only observer based options are updated between renders (like staleTime), but query level ones are not (like retryDelay). This makes no sense, so trying to streamline here

The deleted test checked for multiple useQuery calls in the tree with different options, which I don't think happens a lot in practice.

As a side-effect, the queryFn is also updated, so it will see values from a closure. This shouldn't matter because everything used in the queryFn must be part of the queryKey

* test: solid adapter